+++ nengo_loihi/loihi_interface.py
@@ -32,7 +32,12 @@ except ImportError:
 import nengo_loihi.loihi_cx as loihi_cx
 from nengo_loihi.allocators import one_to_one_allocator
 from nengo_loihi.loihi_api import (
-    bias_to_manexp, CX_PROFILES_MAX, SpikeInput, VTH_PROFILES_MAX)
+    bias_to_manexp,
+    CX_PROFILES_MAX,
+    scale_pes_errors,
+    SpikeInput,
+    VTH_PROFILES_MAX,
+)
 from nengo_loihi.loihi_cx import CxGroup
 
 logger = logging.getLogger(__name__)
@@ -593,6 +598,7 @@ class LoihiSimulator(object):
     def build(self, cx_model, seed=None):
         cx_model.validate()
         self.model = cx_model
+        self.pes_error_scale = getattr(cx_model, 'pes_error_scale', 1.)
 
         if self.use_snips:
             # tag all probes as being snip-based,
@@ -729,8 +735,6 @@ class LoihiSimulator(object):
 
         loihi_errors = []
         for synapses, t, e in errors:
-            x = (100 * e).astype(int)
-            x = np.clip(x, -100, 100, out=x)
             cx_group = synapses.group
             coreid = None
             for core in self.board.chips[0].cores:
@@ -744,7 +748,8 @@ class LoihiSimulator(object):
                     break
 
             assert coreid is not None
-            loihi_errors.append([coreid, len(x)] + x.tolist())
+            e = scale_pes_errors(e, scale=self.pes_error_scale)
+            loihi_errors.append([coreid, len(e)] + e.tolist())
 
         if self.use_snips:
             return self._host2chip_snips(loihi_spikes, loihi_errors)